// apps/api/src/types/omdb.types.ts
// Datos crudos de omdbapi
export interface OMDBMovie {
  Title: string;
  Year: string;
  Rated: string;
  Released: string;
  Runtime: string;
  Genre: string;
  Director: string;
  Writer: string;
  Actors: string;
  Plot: string;
  Language: string;
  Country: string;
  Awards: string;
  Poster: string;
  Ratings: Array<{
    Source: string;
    Value: string;
  }>;
  Metascore: string;
  imdbRating: string;
  imdbVotes: string;
  imdbID: string;
  Type: string;
  DVD: string;
  BoxOffice: string;
  Production: string;
  Website: string;
  Response: string;
}

export interface OMDBSuccessResponse extends OMDBMovie {
  Response: 'True';
}

export interface OMDBErrorResponse {
  Response: 'False';
  Error: string;
}

export type OMDBResponse = OMDBSuccessResponse | OMDBErrorResponse;